Add a factory method that creates dummys.
Gerard Meszaros defined various types of test doubles. One of these
types is the dummy object. It is only passed as a parameter but never
used. Such an object can now be created with the factory method
`Mockito.dummy(AnyType.class)` or by annotation a field with @Dummy. A
dummy throws an exception if any of its methods is called. This is
helpful because the dummy object is no longer a dummy object if this
happens and that means an expectation about the SUT is wrong.